pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atenea"
version = "0.1.0"
description = "Adaptive learning from documents — knowledge extraction, MIR question generation, spaced repetition"
requires-python = ">=3.10"
authors = [
    {name = "Yan Rodriguez Hachimaru"},
]

dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "litellm>=1.0",
    "langdetect>=1.0",
    "pdfplumber>=0.10",
    "PyMuPDF>=1.24",
    "python-dotenv>=1.0",
    "pyzotero>=1.5",
    "nicegui>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
native = ["pywebview>=5.0"]

[project.scripts]
atenea = "atenea.cli:main"
atenea-web = "atenea.web.app:start"

[tool.setuptools.packages.find]
include = ["atenea*", "config*"]
